+++ src/Mod/Sketcher/Gui/CommandConstraints.cpp
@@ -48,6 +48,8 @@
 #include <Mod/Sketcher/App/Sketch.h>
 #include <Mod/Sketcher/App/GeometryFacade.h>
 
+#include "Utils.h"
+
 #include "ViewProviderSketch.h"
 #include "DrawSketchHandler.h"
 #include "ui_InsertDatum.h"
@@ -161,159 +163,6 @@ void showNoConstraintBetweenFixedGeometry()
                                      "Fixed geometries involve external geometry, "
                                      "blocked geometry or special points "
                                      "as B-spline knot points."));
-}
-
-bool SketcherGui::checkBothExternal(int GeoId1, int GeoId2)
-{
-    if (GeoId1 == Constraint::GeoUndef || GeoId2 == Constraint::GeoUndef)
-        return false;
-    else
-        return (GeoId1 < 0 && GeoId2 < 0);
-}
-
-bool SketcherGui::checkBothExternalOrBSplinePoints(const Sketcher::SketchObject* Obj,int GeoId1, int GeoId2)
-{
-    if (GeoId1 == Constraint::GeoUndef || GeoId2 == Constraint::GeoUndef)
-        return false;
-    else
-        return (GeoId1 < 0 && GeoId2 < 0) || (isBsplineKnot(Obj,GeoId1) && isBsplineKnot(Obj,GeoId2)) ||
-        (GeoId1 < 0 && isBsplineKnot(Obj,GeoId2)) || (GeoId2 < 0 && isBsplineKnot(Obj,GeoId1));
-}
-
-bool SketcherGui::isPointOrSegmentFixed(const Sketcher::SketchObject* Obj, int GeoId)
-{
-    const std::vector< Sketcher::Constraint * > &vals = Obj->Constraints.getValues();
-
-    if (GeoId == Constraint::GeoUndef)
-        return false;
-    else
-        return checkConstraint(vals, Sketcher::Block, GeoId, Sketcher::none) || GeoId <= Sketcher::GeoEnum::RtPnt || isBsplineKnot(Obj,GeoId);
-}
-
-bool SketcherGui::areBothPointsOrSegmentsFixed(const Sketcher::SketchObject* Obj, int GeoId1, int GeoId2)
-{
-    const std::vector< Sketcher::Constraint * > &vals = Obj->Constraints.getValues();
-
-    if (GeoId1 == Constraint::GeoUndef || GeoId2 == Constraint::GeoUndef)
-        return false;
-    else
-        return ((checkConstraint(vals, Sketcher::Block, GeoId1, Sketcher::none) || GeoId1 <= Sketcher::GeoEnum::RtPnt || isBsplineKnot(Obj,GeoId1)) &&
-                (checkConstraint(vals, Sketcher::Block, GeoId2, Sketcher::none) || GeoId2 <= Sketcher::GeoEnum::RtPnt || isBsplineKnot(Obj,GeoId2)));
-}
-
-bool SketcherGui::areAllPointsOrSegmentsFixed(const Sketcher::SketchObject* Obj, int GeoId1, int GeoId2, int GeoId3)
-{
-    const std::vector< Sketcher::Constraint * > &vals = Obj->Constraints.getValues();
-
-    if (GeoId1 == Constraint::GeoUndef || GeoId2 == Constraint::GeoUndef || GeoId3 == Constraint::GeoUndef)
-        return false;
-    else
-        return ((checkConstraint(vals, Sketcher::Block, GeoId1, Sketcher::none) || GeoId1 <= Sketcher::GeoEnum::RtPnt || isBsplineKnot(Obj,GeoId1)) &&
-                (checkConstraint(vals, Sketcher::Block, GeoId2, Sketcher::none) || GeoId2 <= Sketcher::GeoEnum::RtPnt || isBsplineKnot(Obj,GeoId2)) &&
-                (checkConstraint(vals, Sketcher::Block, GeoId3, Sketcher::none) || GeoId3 <= Sketcher::GeoEnum::RtPnt || isBsplineKnot(Obj,GeoId3)));
-}
-
-void SketcherGui::getIdsFromName(const std::string &name, const Sketcher::SketchObject* Obj,
-                                 int &GeoId, PointPos &PosId)
-{
-    GeoId = Constraint::GeoUndef;
-    PosId = Sketcher::none;
-
-    if (name.size() > 4 && name.substr(0,4) == "Edge") {
-        GeoId = std::atoi(name.substr(4,4000).c_str()) - 1;
-    }
-    else if (name.size() == 9 && name.substr(0,9) == "RootPoint") {
-        GeoId = Sketcher::GeoEnum::RtPnt;
-        PosId = Sketcher::start;
-    }
-    else if (name.size() == 6 && name.substr(0,6) == "H_Axis")
-        GeoId = Sketcher::GeoEnum::HAxis;
-    else if (name.size() == 6 && name.substr(0,6) == "V_Axis")
-        GeoId = Sketcher::GeoEnum::VAxis;
-    else if (name.size() > 12 && name.substr(0,12) == "ExternalEdge")
-        GeoId = Sketcher::GeoEnum::RefExt + 1 - std::atoi(name.substr(12,4000).c_str());
-    else if (name.size() > 6 && name.substr(0,6) == "Vertex") {
-        int VtId = std::atoi(name.substr(6,4000).c_str()) - 1;
-        Obj->getGeoVertexIndex(VtId,GeoId,PosId);
-    }
-}
-
-bool inline SketcherGui::isVertex(int GeoId, PointPos PosId)
-{
-    return (GeoId != Constraint::GeoUndef && PosId != Sketcher::none);
-}
-
-bool inline SketcherGui::isEdge(int GeoId, PointPos PosId)
-{
-    return (GeoId != Constraint::GeoUndef && PosId == Sketcher::none);
-}
-
-bool SketcherGui::isSimpleVertex(const Sketcher::SketchObject* Obj, int GeoId, PointPos PosId)
-{
-    if (PosId == Sketcher::start && (GeoId == Sketcher::GeoEnum::HAxis || GeoId == Sketcher::GeoEnum::VAxis))
-        return true;
-    const Part::Geometry *geo = Obj->getGeometry(GeoId);
-    if (geo->getTypeId() == Part::GeomPoint::getClassTypeId())
-        return true;
-    else if (PosId == Sketcher::mid)
-        return true;
-    else
-        return false;
-}
-
-bool SketcherGui::isBsplineKnot(const Sketcher::SketchObject* Obj, int GeoId)
-{
-    auto gf = Obj->getGeometryFacade(GeoId);
-    return (gf && gf->getInternalType() == Sketcher::InternalType::BSplineKnotPoint);
-}
-
-bool SketcherGui::IsPointAlreadyOnCurve(int GeoIdCurve, int GeoIdPoint, Sketcher::PointPos PosIdPoint, Sketcher::SketchObject* Obj)
-{
-    //This func is a "smartness" behind three-element tangent-, perp.- and angle-via-point.
-    //We want to find out, if the point supplied by user is already on
-    // both of the curves. If not, necessary point-on-object constraints
-    // are to be added automatically.
-    //Simple geometric test seems to be the best, because a point can be
-    // constrained to a curve in a number of ways (e.g. it is an endpoint of an
-    // arc, or is coincident to endpoint of an arc, or it is an endpoint of an
-    // ellipse's majopr diameter line). Testing all those possibilities is way
-    // too much trouble, IMO(DeepSOIC).
-    Base::Vector3d p = Obj->getPoint(GeoIdPoint, PosIdPoint);
-    return Obj->isPointOnCurve(GeoIdCurve, p.x, p.y);
-}
-
-bool SketcherGui::isBsplinePole(const Part::Geometry * geo)
-{
-    auto gf = GeometryFacade::getFacade(geo);
-
-    if(gf)
-        return gf->getInternalType() == InternalType::BSplineControlPoint;
-
-    THROWM(Base::ValueError, "Null geometry in isBsplinePole - please report")
-}
-
-bool SketcherGui::isBsplinePole(const Sketcher::SketchObject* Obj, int GeoId)
-{
-
-    auto geom = Obj->getGeometry(GeoId);
-
-    return isBsplinePole(geom);
-}
-
-bool SketcherGui::ReleaseHandler(Gui::Document* doc) {
-    if (doc) {
-        if (doc->getInEdit() && doc->getInEdit()->isDerivedFrom(SketcherGui::ViewProviderSketch::getClassTypeId())) {
-            SketcherGui::ViewProviderSketch* vp = static_cast<SketcherGui::ViewProviderSketch*> (doc->getInEdit());
-
-            if (static_cast<SketcherGui::ViewProviderSketch*>(doc->getInEdit())
-                ->getSketchMode() == ViewProviderSketch::STATUS_SKETCH_UseHandler) {
-
-                vp->purgeHandler();
-                return true;
-            }
-        }
-    }
-    return false;
 }
 
 /// Makes a simple tangency constraint using extra point + tangent via point
@@ -601,70 +450,6 @@ void SketcherGui::makeTangentToArcOfParabolaviaNewPoint(Sketcher::SketchObject* 
 
     Gui::Command::commitCommand();
     tryAutoRecompute(Obj);
-}
-
-std::string SketcherGui::getStrippedPythonExceptionString(const Base::Exception& e)
-{
-    std::string msg = e.what();
-
-    if( msg.length() > 26 && msg.substr(0,26) == "FreeCAD exception thrown (") {
-        return msg.substr(26, msg.length()-27);
-    }
-    else
-        return msg;
-}
-
-bool SketcherGui::tryAutoRecompute(Sketcher::SketchObject* obj, bool &autoremoveredundants)
-{
-    ParameterGrp::handle hGrp = App::GetApplication().GetParameterGroupByPath("User parameter:BaseApp/Preferences/Mod/Sketcher");
-    bool autoRecompute = hGrp->GetBool("AutoRecompute",false);
-    bool autoRemoveRedundants = hGrp->GetBool("AutoRemoveRedundants",false);
-
-    // We need to make sure the solver has right redundancy information before trying to remove the redundants.
-    // for example if a non-driving constraint has been added.
-    if(autoRemoveRedundants && autoRecompute)
-        obj->solve();
-
-    if(autoRemoveRedundants)
-        obj->autoRemoveRedundants();
-
-    if (autoRecompute)
-        Gui::Command::updateActive();
-
-    autoremoveredundants = autoRemoveRedundants;
-
-    return autoRecompute;
-}
-
-bool SketcherGui::tryAutoRecompute(Sketcher::SketchObject* obj)
-{
-    bool autoremoveredundants;
-
-    return tryAutoRecompute(obj,autoremoveredundants);
-}
-
-void SketcherGui::tryAutoRecomputeIfNotSolve(Sketcher::SketchObject* obj)
-{
-    bool autoremoveredundants;
-
-    if(!tryAutoRecompute(obj,autoremoveredundants)) {
-        obj->solve();
-
-        if(autoremoveredundants) {
-            obj->autoRemoveRedundants();
-        }
-    }
-}
-
-bool SketcherGui::checkConstraint(const std::vector< Sketcher::Constraint * > &vals, ConstraintType type, int geoid, PointPos pos)
-{
-    for (std::vector< Sketcher::Constraint * >::const_iterator itc= vals.begin(); itc != vals.end(); ++itc) {
-        if ((*itc)->Type == type && (*itc)->First == geoid && (*itc)->FirstPos == pos){
-            return true;
-        }
-    }
-
-    return false;
 }
 
 void SketcherGui::doEndpointTangency(Sketcher::SketchObject* Obj,
